requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
threadpoolctl==3.5.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
threadpoolctl==3.5.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.1
apache-airflow-providers-sqlite==3.9.1
apispec==6.7.1
argcomplete==3.5.1
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
constantly==23.10.4
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==43.0.3
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==24.7.2
inflection==0.5.1
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
numpy==2.1.3
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pathspec==0.12.1
pendulum==3.0.0
pillow==11.0.0
pluggy==1.5.0
prison==0.2.1
propcache==0.2.0
Protego==0.3.1
protobuf==5.28.3
psutil==6.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.10.0
pymongo==4.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.12.0
selenium==4.26.1
service-identity==24.2.0
setproctitle==1.3.4
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.2
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tldextract==5.1.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.5
urllib3==2.2.3
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.17.2
zipp==3.21.0
zope.interface==7.1.1
